rpc: initial implementation
RPC package contains end to end typesafe server and client implementations.

Basic idea of this package is to provide an easy to use API with end to
end typesafe function calls between client and server.

The typesafety at this point only applies to typescript compiler. If used
in JavaScript environments - type safety is not expected, as there is
no runtime type validation implemented.

For example schema definition in server-side setting a query for
`getUsers: () => User[]`, will correctly be intepreted in the client side
to receive type of User as an array.

In this commit the basic functionality is introduced.
Queries and mutations can be made succesfully and customised with end to
end typesafety. (Disclaimer: only tsc - no error handling or runtime
type validation)

Subscriptions (websocket) works only with the connection part. No
event-streaming has been implemented yet.